Make a global Contact support widget that listens to URL hash fragment changes which triggers contact support form.
