+++ client/a8c-for-agencies/components/a4a-migration-offer/index.tsx
@@ -1,8 +1,7 @@
-import page from '@automattic/calypso-router';
 import { Button, FoldableCard } from '@automattic/components';
 import { Icon, reusableBlock, external } from '@wordpress/icons';
 import { useTranslate } from 'i18n-calypso';
-import { A4A_OVERVIEW_LINK } from '../sidebar-menu/lib/constants';
+import { CONTACT_URL_FOR_MIGRATION_OFFER_HASH_FRAGMENT } from '../a4a-contact-support-widget';
 
 import './style.scss';
 type Props = {
@@ -29,12 +28,7 @@ const MigrationOfferBody = () => {
 	return (
 		<>
 			<p className="a4a-migration-offer__description">{ description }</p>
-			<Button
-				onClick={ () => {
-					page( `${ A4A_OVERVIEW_LINK }#contact-support-migration-offer` );
-				} }
-				primary
-			>
+			<Button href={ CONTACT_URL_FOR_MIGRATION_OFFER_HASH_FRAGMENT } primary>
 				{ translate( 'Contact us' ) }
 				<Icon icon={ external } size={ 18 } />
 			</Button>
