+++ client/a8c-for-agencies/components/sidebar/header/profile-dropdown.tsx
@@ -1,18 +1,16 @@
-import page from '@automattic/calypso-router';
 import { Button, Gridicon, Gravatar } from '@automattic/components';
 import { Icon, chevronDown } from '@wordpress/icons';
 import clsx from 'clsx';
 import { useTranslate } from 'i18n-calypso';
 import { useCallback, useRef, useState } from 'react';
-import { CONTACT_URL_HASH_FRAGMENT } from 'calypso/a8c-for-agencies/sections/overview/sidebar/contact-support';
 import { isClientView } from 'calypso/a8c-for-agencies/sections/purchases/payment-methods/lib/is-client-view';
 import useOutsideClickCallback from 'calypso/lib/use-outside-click-callback';
 import { useDispatch, useSelector } from 'calypso/state';
 import { recordTracksEvent } from 'calypso/state/analytics/actions/record';
 import { redirectToLogout } from 'calypso/state/current-user/actions';
 import { getCurrentUser } from 'calypso/state/current-user/selectors';
+import { CONTACT_URL_HASH_FRAGMENT } from '../../a4a-contact-support-widget';
 import {
-	A4A_OVERVIEW_LINK,
 	EXTERNAL_A4A_KNOWLEDGE_BASE,
 	EXTERNAL_A4A_CLIENT_KNOWLEDGE_BASE,
 } from '../../sidebar-menu/lib/constants';
@@ -28,7 +26,6 @@ const DropdownMenu = ( { isExpanded, setMenuExpanded }: DropdownMenuProps ) => {
 	const translate = useTranslate();
 
 	const onGetHelp = useCallback( () => {
-		page( A4A_OVERVIEW_LINK + CONTACT_URL_HASH_FRAGMENT );
 		setMenuExpanded( false );
 		dispatch( recordTracksEvent( 'calypso_a4a_sidebar_gethelp' ) );
 	}, [ dispatch, setMenuExpanded ] );
@@ -45,7 +42,7 @@ const DropdownMenu = ( { isExpanded, setMenuExpanded }: DropdownMenuProps ) => {
 				// Show the "Contact support" button if the user is not a client
 				! isClient && (
 					<li className="a4a-sidebar__profile-dropdown-menu-item">
-						<Button borderless onClick={ onGetHelp }>
+						<Button borderless onClick={ onGetHelp } href={ CONTACT_URL_HASH_FRAGMENT }>
 							{ translate( 'Contact support' ) }
 						</Button>
 					</li>
