+++ client/a8c-for-agencies/components/sidebar/index.tsx
@@ -2,9 +2,7 @@ import page from '@automattic/calypso-router';
 import { Icon, starEmpty } from '@wordpress/icons';
 import clsx from 'clsx';
 import { useTranslate } from 'i18n-calypso';
-import { useCallback, useState } from 'react';
-import UserContactSupportModalForm from 'calypso/a8c-for-agencies/components/user-contact-support-modal-form';
-import { CONTACT_URL_HASH_FRAGMENT } from 'calypso/a8c-for-agencies/sections/overview/sidebar/contact-support';
+import { useCallback } from 'react';
 import { isClientView } from 'calypso/a8c-for-agencies/sections/purchases/payment-methods/lib/is-client-view';
 import JetpackIcons from 'calypso/components/jetpack/sidebar/menu-items/jetpack-icons';
 import Sidebar, {
@@ -16,7 +14,7 @@ import Sidebar, {
 } from 'calypso/layout/sidebar-v2';
 import { useDispatch } from 'calypso/state';
 import { recordTracksEvent } from 'calypso/state/analytics/actions';
-import { A4A_OVERVIEW_LINK, A4A_CLIENT_SUBSCRIPTIONS_LINK } from '../sidebar-menu/lib/constants';
+import A4AContactSupportWidget, { CONTACT_URL_HASH_FRAGMENT } from '../a4a-contact-support-widget';
 import SidebarHeader from './header';
 import ProfileDropdown from './header/profile-dropdown';
 
@@ -63,22 +61,10 @@ const A4ASidebar = ( {
 
 	const isClient = isClientView();
 
-	const [ showUserSupportForm, setShowUserSupportForm ] = useState( false );
-
 	const onShowUserSupportForm = useCallback( () => {
 		dispatch( recordTracksEvent( 'calypso_jetpack_sidebar_share_product_feedback_click' ) );
-		setShowUserSupportForm( true );
 	}, [ dispatch ] );
 
-	const onCloseUserSupportForm = useCallback( () => {
-		// Remove any hash from the URL.
-		history.pushState( null, '', window.location.pathname + window.location.search );
-		setShowUserSupportForm( false );
-	}, [] );
-
-	const contactUsUrl = isClient
-		? A4A_CLIENT_SUBSCRIPTIONS_LINK + '#contact-support'
-		: A4A_OVERVIEW_LINK + '#contact-sales';
 	const contactUsText = isClient ? translate( 'Contact support' ) : translate( 'Contact sales' );
 
 	return (
@@ -121,7 +107,7 @@ const A4ASidebar = ( {
 							path=""
 							icon={ <JetpackIcons icon="help" /> }
 							onClickMenuItem={ () => {
-								page( A4A_OVERVIEW_LINK + CONTACT_URL_HASH_FRAGMENT );
+								page( CONTACT_URL_HASH_FRAGMENT );
 								dispatch(
 									recordTracksEvent( 'calypso_a4a_sidebar_menu_click', {
 										menu_item: 'A4A / Support',
@@ -133,7 +119,7 @@ const A4ASidebar = ( {
 
 					<SidebarNavigatorMenuItem
 						title={ contactUsText }
-						link={ contactUsUrl }
+						link={ CONTACT_URL_HASH_FRAGMENT }
 						path=""
 						icon={ <Icon icon={ starEmpty } /> }
 						onClickMenuItem={ onShowUserSupportForm }
@@ -143,10 +129,7 @@ const A4ASidebar = ( {
 				</ul>
 			</SidebarFooter>
 
-			<UserContactSupportModalForm
-				show={ showUserSupportForm }
-				onClose={ onCloseUserSupportForm }
-			/>
+			<A4AContactSupportWidget />
 		</Sidebar>
 	);
 };
